baseobjspace.rs: split descroperation + formatting into upstream-aligned modules
baseobjspace.rs (8816→6083 lines) fused 8+ upstream PyPy modules.
Extract two new files matching upstream boundaries:

- objspace/descroperation.rs (2184 lines) ↔ pypy/objspace/descroperation.py
  Binary/unary op dispatch (add/sub/mul/compare/pos/neg/…),
  type-specific arithmetic fast paths, CompareOp enum, float_pow_raw.

- objspace/std/formatting.rs (578 lines) ↔ pypy/objspace/std/formatting.py
  str_format_percent, format_g_like, normalise_exponent.

baseobjspace.rs re-exports via `pub use descroperation::*` so all
existing crate::baseobjspace::add / CompareOp paths resolve unchanged.
