+++ pyre/pyre-interpreter/src/baseobjspace.rs
@@ -1,9 +1,12 @@
 //! ObjSpace — Python object operation dispatch.
 #![allow(non_camel_case_types, non_snake_case)]
 //!
-//! The ObjSpace mediates all operations on Python objects. This is the layer
-//! where type-specific fast paths live, and where the JIT inserts `GuardClass`
-//! to specialize operations.
+//! pypy/interpreter/baseobjspace.py — the core ObjSpace interface.
+//!
+//! Binary/unary operation dispatch (add, sub, compare, etc.) lives in
+//! `crate::objspace::descroperation`; printf-style formatting lives in
+//! `crate::objspace::std::formatting`. Both are re-exported here so
+//! existing `crate::baseobjspace::add` paths continue to resolve.
 
 // Suppress unsafe-in-unsafe-fn warnings; our unsafe fns are inherently
 // working with raw pointers throughout and wrapping every call in an
@@ -11,8 +14,6 @@
 #![allow(unsafe_op_in_unsafe_fn)]
 
 use malachite_bigint::BigInt;
-use num_integer::Integer;
-use num_traits::ToPrimitive;
 
 use std::cell::RefCell;
 use std::collections::HashMap;
@@ -21,6 +22,11 @@ use crate::function::is_function;
 pub use crate::{PyError, PyErrorKind, PyResult};
 use pyre_object::strobject::is_str;
 use pyre_object::*;
+
+// ── Re-exports from split-out modules ────────────────────────────────
+pub use crate::objspace::descroperation::*;
+pub(crate) use crate::objspace::std::formatting::{format_g_like, normalise_exponent};
+
 
 /// Compatibility alias for PyPy's base-object type.
 /// PyPy frequently models interpreter values as subclasses of `W_Root`.
@@ -145,1768 +151,6 @@ pub fn unwrap_cell(obj: PyObjectRef) -> PyObjectRef {
     obj
 }
 
-// ── BigInt helpers ──────────────────────────────────────────────────
-
-/// Extract a BigInt from an int or long object.
-
-unsafe fn as_bigint(obj: PyObjectRef) -> BigInt {
-    if is_int(obj) {
-        BigInt::from(w_int_get_value(obj))
-    } else if is_bool(obj) {
-        BigInt::from(w_bool_get_value(obj) as i64)
-    } else {
-        w_long_get_value(obj).clone()
-    }
-}
-
-/// Box a BigInt result, demoting to W_IntObject if it fits in i64.
-
-fn bigint_result(value: BigInt) -> PyObjectRef {
-    match value.to_i64() {
-        Some(v) => w_int_new(v),
-        None => w_long_new(value),
-    }
-}
-
-#[majit_macros::elidable]
-fn bigint_add(a: BigInt, b: BigInt) -> BigInt {
-    a + b
-}
-
-#[majit_macros::elidable]
-fn bigint_sub(a: BigInt, b: BigInt) -> BigInt {
-    a - b
-}
-
-#[majit_macros::elidable]
-fn bigint_mul(a: BigInt, b: BigInt) -> BigInt {
-    a * b
-}
-
-#[majit_macros::elidable]
-fn bigint_and(a: BigInt, b: BigInt) -> BigInt {
-    a & b
-}
-
-#[majit_macros::elidable]
-fn bigint_or(a: BigInt, b: BigInt) -> BigInt {
-    a | b
-}
-
-#[majit_macros::elidable]
-fn bigint_xor(a: BigInt, b: BigInt) -> BigInt {
-    a ^ b
-}
-
-#[majit_macros::elidable]
-fn bigint_lshift(a: BigInt, shift: usize) -> BigInt {
-    a << shift
-}
-
-#[majit_macros::elidable]
-fn bigint_rshift(a: BigInt, shift: usize) -> BigInt {
-    a >> shift
-}
-
-#[majit_macros::elidable]
-fn bigint_neg(a: BigInt) -> BigInt {
-    -a
-}
-
-#[majit_macros::elidable]
-fn bigint_invert(a: BigInt) -> BigInt {
-    !a
-}
-
-#[majit_macros::elidable]
-fn bigint_eq(a: BigInt, b: BigInt) -> bool {
-    a == b
-}
-
-#[majit_macros::elidable]
-fn bigint_lt(a: BigInt, b: BigInt) -> bool {
-    a < b
-}
-
-#[majit_macros::elidable]
-fn bigint_gt(a: BigInt, b: BigInt) -> bool {
-    a > b
-}
-
-#[majit_macros::elidable]
-fn bigint_mod(a: BigInt, b: BigInt) -> BigInt {
-    a % b
-}
-
-// ── Arithmetic operations ─────────────────────────────────────────────
-
-/// Integer addition fast path.
-///
-/// The JIT will specialize this via:
-///   GuardClass(a, &INT_TYPE)
-///   GuardClass(b, &INT_TYPE)
-///   GetfieldGcI(a, intval_offset) → va
-///   GetfieldGcI(b, intval_offset) → vb
-///   IntAdd(va, vb) → result
-///   New(W_IntObject) + SetfieldGcI(result)
-
-unsafe fn int_add(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    match va.checked_add(vb) {
-        Some(r) => Ok(w_int_new(r)),
-        None => Ok(w_long_new(bigint_add(BigInt::from(va), BigInt::from(vb)))),
-    }
-}
-
-unsafe fn int_sub(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    match va.checked_sub(vb) {
-        Some(r) => Ok(w_int_new(r)),
-        None => Ok(w_long_new(bigint_sub(BigInt::from(va), BigInt::from(vb)))),
-    }
-}
-
-unsafe fn int_mul(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    match va.checked_mul(vb) {
-        Some(r) => Ok(w_int_new(r)),
-        None => Ok(w_long_new(bigint_mul(BigInt::from(va), BigInt::from(vb)))),
-    }
-}
-
-unsafe fn int_floordiv(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    if vb == 0 {
-        return Err(PyError::zero_division("integer division or modulo by zero"));
-    }
-    // Python floor division: rounds toward negative infinity.
-    // i64::MIN / -1 overflows → fall back to BigInt.
-    let q = match va.checked_div(vb) {
-        Some(q) => q,
-        None => return Ok(bigint_result(BigInt::from(va).div_floor(&BigInt::from(vb)))),
-    };
-    let r = va % vb;
-    // Adjust: if remainder is nonzero and signs of operands differ, subtract 1.
-    let q = if r != 0 && (r ^ vb) < 0 { q - 1 } else { q };
-    Ok(w_int_new(q))
-}
-
-unsafe fn int_mod(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    if vb == 0 {
-        return Err(PyError::zero_division("integer division or modulo by zero"));
-    }
-    // Python modulo: result has the same sign as the divisor.
-    let r = va % vb;
-    let r = if r != 0 && (r ^ vb) < 0 { r + vb } else { r };
-    Ok(w_int_new(r))
-}
-
-// ── Long (BigInt) arithmetic operations ─────────────────────────────
-
-unsafe fn long_add(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(bigint_result(bigint_add(as_bigint(a), as_bigint(b))))
-}
-
-unsafe fn long_sub(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(bigint_result(bigint_sub(as_bigint(a), as_bigint(b))))
-}
-
-unsafe fn long_mul(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(bigint_result(bigint_mul(as_bigint(a), as_bigint(b))))
-}
-
-unsafe fn long_floordiv(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let vb = as_bigint(b);
-    if bigint_eq(vb.clone(), BigInt::from(0)) {
-        return Err(PyError::zero_division("integer division or modulo by zero"));
-    }
-    Ok(bigint_result(as_bigint(a).div_floor(&vb)))
-}
-
-unsafe fn long_mod(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let vb = as_bigint(b);
-    if bigint_eq(vb.clone(), BigInt::from(0)) {
-        return Err(PyError::zero_division("integer division or modulo by zero"));
-    }
-    Ok(bigint_result(as_bigint(a).mod_floor(&vb)))
-}
-
-// ── Float arithmetic operations ──────────────────────────────────────
-
-/// Coerce an operand to f64. Works for int, long, and float objects.
-unsafe fn as_float(obj: PyObjectRef) -> f64 {
-    if is_float(obj) {
-        w_float_get_value(obj)
-    } else if is_int(obj) {
-        w_int_get_value(obj) as f64
-    } else {
-        // long → f64 (may lose precision for very large values)
-        w_long_get_value(obj).to_f64().unwrap_or(f64::INFINITY)
-    }
-}
-
-/// True if both operands are numeric and at least one is float.
-
-unsafe fn is_float_pair(a: PyObjectRef, b: PyObjectRef) -> bool {
-    let a_num = is_int(a) || is_float(a) || is_long(a);
-    let b_num = is_int(b) || is_float(b) || is_long(b);
-    a_num && b_num && (is_float(a) || is_float(b))
-}
-
-unsafe fn float_add(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_float_new(as_float(a) + as_float(b)))
-}
-
-unsafe fn float_sub(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_float_new(as_float(a) - as_float(b)))
-}
-
-unsafe fn float_mul(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_float_new(as_float(a) * as_float(b)))
-}
-
-unsafe fn float_truediv(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let vb = as_float(b);
-    if vb == 0.0 {
-        return Err(PyError::zero_division("float division by zero"));
-    }
-    Ok(w_float_new(as_float(a) / vb))
-}
-
-/// floatobject.py:508-512: descr_floordiv → _divmod_w()[0].
-unsafe fn float_floordiv(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let (floordiv, _mod) = float_divmod_w(as_float(a), as_float(b))?;
-    Ok(w_float_new(floordiv))
-}
-
-/// floatobject.py:520-540: descr_mod with math_fmod + sign correction.
-unsafe fn float_mod(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let x = as_float(a);
-    let y = as_float(b);
-    if y == 0.0 {
-        // floatobject.py:526
-        return Err(PyError::zero_division("float modulo"));
-    }
-    let mut m = x % y; // fmod
-    if m != 0.0 {
-        // floatobject.py:529-531: ensure remainder has same sign as denominator
-        if (y < 0.0) != (m < 0.0) {
-            m += y;
-        }
-    } else {
-        // floatobject.py:536-538: signed zero — copysign(0.0, y)
-        m = f64::copysign(0.0, y);
-    }
-    Ok(w_float_new(m))
-}
-
-/// floatobject.py:758-793: _divmod_w.
-fn float_divmod_w(x: f64, y: f64) -> Result<(f64, f64), PyError> {
-    if y == 0.0 {
-        // floatobject.py:761
-        return Err(PyError::zero_division("float modulo"));
-    }
-    let mut m = x % y; // fmod
-    // floatobject.py:767: div = (x - mod) / y
-    let mut div = (x - m) / y;
-    if m != 0.0 {
-        // floatobject.py:769-771: sign correction
-        if (y < 0.0) != (m < 0.0) {
-            m += y;
-            div -= 1.0;
-        }
-    } else {
-        // floatobject.py:776-778: signed zero
-        // "mod *= mod" hides "+0" from optimizer, then negate if y < 0
-        m = m * m; // hide from optimizer
-        if y < 0.0 {
-            m = -m;
-        }
-    }
-    // floatobject.py:784-790: snap quotient to nearest integral value
-    let floordiv = if div != 0.0 {
-        let f = div.floor();
-        if div - f > 0.5 { f + 1.0 } else { f }
-    } else {
-        // floatobject.py:789-790: zero with sign of true quotient
-        let d = div * div; // hide from optimizer
-        d * x / y
-    };
-    Ok((floordiv, m))
-}
-
-// ── Power ────────────────────────────────────────────────────────────
-
-unsafe fn int_pow(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    if vb < 0 {
-        // Negative exponent → float result
-        return Ok(w_float_new((va as f64).powf(vb as f64)));
-    }
-    let vb = vb as u32;
-    match va.checked_pow(vb) {
-        Some(r) => Ok(w_int_new(r)),
-        None => Ok(w_long_new(BigInt::from(va).pow(vb))),
-    }
-}
-
-unsafe fn long_pow(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let vb = as_bigint(b);
-    if bigint_lt(vb.clone(), BigInt::from(0)) {
-        let fa = as_float(a);
-        let fb = as_float(b);
-        return Ok(w_float_new(fa.powf(fb)));
-    }
-    let exp = vb.to_u32().unwrap_or(u32::MAX);
-    Ok(bigint_result(as_bigint(a).pow(exp)))
-}
-
-// ── Shift operations ─────────────────────────────────────────────────
-
-unsafe fn int_lshift(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    if vb < 0 {
-        return Err(PyError::value_error("negative shift count"));
-    }
-    // `i64::checked_shl` only fails when the shift amount is >= 64, so it
-    // happily returns a wrapped result when the VALUE overflows (e.g.
-    // `(10**18) << 4`). Detect real value overflow by computing the shift
-    // in BigInt and demoting to i64 only when the result fits.
-    let big = bigint_lshift(BigInt::from(va), vb as usize);
-    Ok(bigint_result(big))
-}
-
-unsafe fn int_rshift(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let va = int_value(a);
-    let vb = int_value(b);
-    if vb < 0 {
-        return Err(PyError::value_error("negative shift count"));
-    }
-    let vb = vb as u32;
-    Ok(w_int_new(va >> vb.min(63)))
-}
-
-unsafe fn long_lshift(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let vb = as_bigint(b);
-    if bigint_lt(vb.clone(), BigInt::from(0)) {
-        return Err(PyError::value_error("negative shift count"));
-    }
-    let shift = vb.to_u32().unwrap_or(u32::MAX);
-    Ok(bigint_result(bigint_lshift(as_bigint(a), shift as usize)))
-}
-
-unsafe fn long_rshift(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let vb = as_bigint(b);
-    if bigint_lt(vb.clone(), BigInt::from(0)) {
-        return Err(PyError::value_error("negative shift count"));
-    }
-    let shift = vb.to_u32().unwrap_or(u32::MAX);
-    Ok(bigint_result(bigint_rshift(as_bigint(a), shift as usize)))
-}
-
-// ── bool-as-int helpers ──────────────────────────────────────────────
-
-/// True when obj is int or bool (bool is a subclass of int in Python).
-#[inline]
-unsafe fn is_int_like(obj: PyObjectRef) -> bool {
-    is_int(obj) || is_bool(obj)
-}
-
-/// Extract i64 from an int or bool object.
-#[inline]
-unsafe fn int_value(obj: PyObjectRef) -> i64 {
-    if is_bool(obj) {
-        w_bool_get_value(obj) as i64
-    } else {
-        w_int_get_value(obj)
-    }
-}
-
-// ── Bitwise operations ───────────────────────────────────────────────
-
-unsafe fn int_bitand(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let r = int_value(a) & int_value(b);
-    // bool & bool → bool
-    if is_bool(a) && is_bool(b) {
-        return Ok(w_bool_from(r != 0));
-    }
-    Ok(w_int_new(r))
-}
-
-unsafe fn int_bitor(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let r = int_value(a) | int_value(b);
-    if is_bool(a) && is_bool(b) {
-        return Ok(w_bool_from(r != 0));
-    }
-    Ok(w_int_new(r))
-}
-
-unsafe fn int_bitxor(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let r = int_value(a) ^ int_value(b);
-    if is_bool(a) && is_bool(b) {
-        return Ok(w_bool_from(r != 0));
-    }
-    Ok(w_int_new(r))
-}
-
-unsafe fn long_bitand(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(bigint_result(bigint_and(as_bigint(a), as_bigint(b))))
-}
-
-unsafe fn long_bitor(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(bigint_result(bigint_or(as_bigint(a), as_bigint(b))))
-}
-
-unsafe fn long_bitxor(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(bigint_result(bigint_xor(as_bigint(a), as_bigint(b))))
-}
-
-// ── String operations ────────────────────────────────────────────────
-
-/// Concatenate two str objects.
-
-unsafe fn str_concat(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let sa = w_str_get_value(a);
-    let sb = w_str_get_value(b);
-    let mut result = String::with_capacity(sa.len() + sb.len());
-    result.push_str(sa);
-    result.push_str(sb);
-    Ok(w_str_new(&result))
-}
-
-/// Repeat a str object `n` times.
-
-unsafe fn str_repeat(s: PyObjectRef, n: PyObjectRef) -> PyResult {
-    let sv = w_str_get_value(s);
-    let nv = w_int_get_value(n);
-    let count = if nv < 0 { 0 } else { nv as usize };
-    Ok(w_str_new(&sv.repeat(count)))
-}
-
-unsafe fn list_concat(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let len_a = w_list_len(a);
-    let len_b = w_list_len(b);
-    let mut items = Vec::with_capacity(len_a + len_b);
-    for i in 0..len_a {
-        if let Some(item) = w_list_getitem(a, i as i64) {
-            items.push(item);
-        }
-    }
-    for i in 0..len_b {
-        if let Some(item) = w_list_getitem(b, i as i64) {
-            items.push(item);
-        }
-    }
-    Ok(w_list_new(items))
-}
-
-unsafe fn tuple_concat(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let len_a = w_tuple_len(a);
-    let len_b = w_tuple_len(b);
-    let mut items = Vec::with_capacity(len_a + len_b);
-    for i in 0..len_a {
-        if let Some(item) = w_tuple_getitem(a, i as i64) {
-            items.push(item);
-        }
-    }
-    for i in 0..len_b {
-        if let Some(item) = w_tuple_getitem(b, i as i64) {
-            items.push(item);
-        }
-    }
-    Ok(w_tuple_new(items))
-}
-
-unsafe fn list_repeat(list: PyObjectRef, n: PyObjectRef) -> PyResult {
-    let nv = w_int_get_value(n);
-    let count = if nv < 0 { 0 } else { nv as usize };
-    let len = w_list_len(list);
-    let mut items = Vec::with_capacity(len * count);
-    for _ in 0..count {
-        for i in 0..len {
-            if let Some(item) = w_list_getitem(list, i as i64) {
-                items.push(item);
-            }
-        }
-    }
-    Ok(w_list_new(items))
-}
-
-// ── Comparison operations ─────────────────────────────────────────────
-
-unsafe fn int_lt(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(int_value(a) < int_value(b)))
-}
-
-unsafe fn int_le(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(int_value(a) <= int_value(b)))
-}
-
-unsafe fn int_gt(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(int_value(a) > int_value(b)))
-}
-
-unsafe fn int_ge(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(int_value(a) >= int_value(b)))
-}
-
-unsafe fn int_eq(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(int_value(a) == int_value(b)))
-}
-
-unsafe fn int_ne(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(int_value(a) != int_value(b)))
-}
-
-unsafe fn float_lt(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(as_float(a) < as_float(b)))
-}
-
-unsafe fn float_le(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(as_float(a) <= as_float(b)))
-}
-
-unsafe fn float_gt(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(as_float(a) > as_float(b)))
-}
-
-unsafe fn float_ge(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(as_float(a) >= as_float(b)))
-}
-
-unsafe fn float_eq(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(as_float(a) == as_float(b)))
-}
-
-unsafe fn float_ne(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    Ok(w_bool_from(as_float(a) != as_float(b)))
-}
-
-// ── Public dispatch API ───────────────────────────────────────────────
-
-/// Try to call a dunder method on an instance for binary ops.
-///
-/// PyPy: descroperation.py `_binop_impl` →
-///   1. Try `a.__op__(b)` (forward)
-///   2. If not found or returns NotImplemented, try `b.__rop__(a)` (reverse)
-///
-/// descroperation.py:432-437 parity: `space.get_and_call_function` raises
-/// OperationError; the NotImplemented return value alone triggers the
-/// fallback. We mirror that by propagating PyError immediately — the
-/// pyre `call_function` shim stashes exceptions in PENDING_CALL_ERROR
-/// so we must consume them via `call_function_impl_result` to match.
-unsafe fn try_instance_binop(a: PyObjectRef, b: PyObjectRef, dunder: &str) -> Option<PyResult> {
-    let a_is_inst = is_instance(a);
-    let b_is_inst = is_instance(b);
-
-    // PyPy: descroperation.py _binop_impl
-    // If b's type is a proper subtype of a's type, try reverse first.
-    // This matches Python's "subclass reflected op takes priority" rule.
-    let try_reverse_first = if a_is_inst && b_is_inst {
-        if let Some(rdunder) = reverse_dunder(dunder) {
-            let a_type = w_instance_get_type(a);
-            let b_type = w_instance_get_type(b);
-            !std::ptr::eq(a_type, b_type)
-                && issubtype_cached(b_type, a_type)
-                && lookup_in_type_where(b_type, rdunder).is_some()
-        } else {
-            false
-        }
-    } else {
-        false
-    };
-
-    if try_reverse_first {
-        let rdunder = reverse_dunder(dunder).unwrap();
-        let w_type = w_instance_get_type(b);
-        if let Some(method) = lookup_in_type_where(w_type, rdunder) {
-            match crate::call::call_function_impl_result(method, &[b, a]) {
-                Ok(result) => {
-                    if !is_not_implemented(result) {
-                        return Some(Ok(result));
-                    }
-                }
-                Err(e) => return Some(Err(e)),
-            }
-        }
-    }
-
-    // Forward: a.__op__(b)
-    if a_is_inst {
-        let w_type = w_instance_get_type(a);
-        if let Some(method) = lookup_in_type_where(w_type, dunder) {
-            match crate::call::call_function_impl_result(method, &[a, b]) {
-                Ok(result) => {
-                    if !is_not_implemented(result) {
-                        return Some(Ok(result));
-                    }
-                }
-                Err(e) => return Some(Err(e)),
-            }
-        }
-        // Also check per-instance attributes (ATTR_TABLE)
-        if let Ok(method) = getattr(a, dunder) {
-            match crate::call::call_function_impl_result(method, &[a, b]) {
-                Ok(result) => {
-                    if !is_not_implemented(result) {
-                        return Some(Ok(result));
-                    }
-                }
-                Err(e) => return Some(Err(e)),
-            }
-        }
-    }
-
-    // Reverse: b.__rop__(a) — only if not already tried above
-    if !try_reverse_first && b_is_inst {
-        if let Some(rdunder) = reverse_dunder(dunder) {
-            let w_type = w_instance_get_type(b);
-            if let Some(method) = lookup_in_type_where(w_type, rdunder) {
-                match crate::call::call_function_impl_result(method, &[b, a]) {
-                    Ok(result) => {
-                        if !is_not_implemented(result) {
-                            return Some(Ok(result));
-                        }
-                    }
-                    Err(e) => return Some(Err(e)),
-                }
-            }
-        }
-    }
-
-    None
-}
-
-/// `descroperation.py _binop_impl` typedef-driven fallback for
-/// non-instance LHS / RHS — pyre's `try_instance_binop` only fires
-/// when at least one operand is `is_instance`, but built-in W_Root
-/// types (dict_view, exception, generator, …) also expose dunder
-/// methods through their typedef.  This helper does the same
-/// forward + reverse MRO lookup but routes through
-/// `crate::typedef::r#type` instead of `w_instance_get_type`, so
-/// `dict_keys() | set()` etc. find the typedef-installed `__or__`
-/// and friends.  Returns `None` when neither side defines the
-/// method (caller falls through to the existing TypeError path).
-unsafe fn try_typedef_binop(a: PyObjectRef, b: PyObjectRef, dunder: &str) -> Option<PyResult> {
-    if let Some(a_type) = crate::typedef::r#type(a) {
-        if let Some(method) = lookup_in_type_where(a_type, dunder) {
-            match crate::call::call_function_impl_result(method, &[a, b]) {
-                Ok(result) => {
-                    if !is_not_implemented(result) {
-                        return Some(Ok(result));
-                    }
-                }
-                Err(e) => return Some(Err(e)),
-            }
-        }
-    }
-    if let Some(rdunder) = reverse_dunder(dunder) {
-        if let Some(b_type) = crate::typedef::r#type(b) {
-            if let Some(method) = lookup_in_type_where(b_type, rdunder) {
-                match crate::call::call_function_impl_result(method, &[b, a]) {
-                    Ok(result) => {
-                        if !is_not_implemented(result) {
-                            return Some(Ok(result));
-                        }
-                    }
-                    Err(e) => return Some(Err(e)),
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Check if w_type is a subtype of cls using cached MRO.
-unsafe fn issubtype_cached(w_type: PyObjectRef, cls: PyObjectRef) -> bool {
-    let mro_ptr = w_type_get_mro(w_type);
-    if !mro_ptr.is_null() {
-        return (*mro_ptr).iter().any(|&t| std::ptr::eq(t, cls));
-    }
-    false
-}
-
-/// Map forward dunder to reverse dunder.
-/// PyPy: descroperation.py `_make_binop_impl` generates both directions.
-fn reverse_dunder(dunder: &str) -> Option<&'static str> {
-    Some(match dunder {
-        // Arithmetic — PyPy: descroperation.py _make_binop_impl
-        "__add__" => "__radd__",
-        "__sub__" => "__rsub__",
-        "__mul__" => "__rmul__",
-        "__truediv__" => "__rtruediv__",
-        "__floordiv__" => "__rfloordiv__",
-        "__mod__" => "__rmod__",
-        "__pow__" => "__rpow__",
-        "__lshift__" => "__rlshift__",
-        "__rshift__" => "__rrshift__",
-        "__and__" => "__rand__",
-        "__or__" => "__ror__",
-        "__xor__" => "__rxor__",
-        // Comparison reflected — PyPy: descroperation.py _cmp_dispatch
-        "__lt__" => "__gt__",
-        "__le__" => "__ge__",
-        "__gt__" => "__lt__",
-        "__ge__" => "__le__",
-        "__eq__" => "__eq__",
-        "__ne__" => "__ne__",
-        _ => return None,
-    })
-}
-
-/// Try to call a unary dunder on an instance.
-///
-/// PyPy: `ObjSpace.call_function(space.lookup(w_obj, dunder), w_obj)`
-/// The Python-level OperationError must propagate to the caller; use the
-/// Result-returning call path so PENDING_CALL_ERROR is consumed.
-unsafe fn try_instance_unaryop(a: PyObjectRef, dunder: &str) -> Option<PyResult> {
-    if is_instance(a) {
-        if let Some(method) = lookup(a, dunder) {
-            return Some(crate::call::call_function_impl_result(method, &[a]));
-        }
-    }
-    None
-}
-
-/// Binary operation dispatch.
-///
-/// Checks types and dispatches to the appropriate fast path.
-/// The JIT traces through this function, recording `GuardClass` on operand types.
-
-pub fn add(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_add(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_add(a, b);
-        }
-        if is_float_pair(a, b) {
-            return float_add(a, b);
-        }
-        if is_str(a) && is_str(b) {
-            return str_concat(a, b);
-        }
-        if is_list(a) && is_list(b) {
-            return list_concat(a, b);
-        }
-        if is_tuple(a) && is_tuple(b) {
-            return tuple_concat(a, b);
-        }
-        if pyre_object::bytesobject::is_bytes_like(a) && pyre_object::bytesobject::is_bytes_like(b)
-        {
-            let a_data = pyre_object::bytesobject::bytes_like_data(a);
-            let b_data = pyre_object::bytesobject::bytes_like_data(b);
-            let mut result = a_data.to_vec();
-            result.extend_from_slice(b_data);
-            // bytes + bytes → bytes; anything with bytearray → bytearray
-            return Ok(
-                if pyre_object::bytesobject::is_bytes(a) && pyre_object::bytesobject::is_bytes(b) {
-                    pyre_object::bytesobject::w_bytes_from_bytes(&result)
-                } else {
-                    pyre_object::bytearrayobject::w_bytearray_from_bytes(&result)
-                },
-            );
-        }
-        // Instance dunder dispatch: __add__
-        if let Some(result) = try_instance_binop(a, b, "__add__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for +: '{}' and '{}'",
-            a_name, b_name,
-        )))
-    }
-}
-
-pub fn sub(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_sub(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_sub(a, b);
-        }
-        if is_float_pair(a, b) {
-            return float_sub(a, b);
-        }
-        // set / frozenset difference — PyPy: setobject.py W_BaseSetObject.descr_sub
-        if pyre_object::is_set_or_frozenset(a) {
-            let other_items = crate::builtins::collect_iterable(b)?;
-            let probe = pyre_object::w_set_from_items(&other_items);
-            let result: Vec<PyObjectRef> = pyre_object::w_set_items(a)
-                .into_iter()
-                .filter(|&item| !pyre_object::w_set_contains(probe, item))
-                .collect();
-            return Ok(if pyre_object::is_frozenset(a) {
-                pyre_object::w_frozenset_from_items(&result)
-            } else {
-                pyre_object::w_set_from_items(&result)
-            });
-        }
-        if let Some(result) = try_instance_binop(a, b, "__sub__") {
-            return result;
-        }
-        // Built-in W_Root types (dict_view, …) expose `__sub__` /
-        // `__rsub__` through their typedef.
-        if let Some(result) = try_typedef_binop(a, b, "__sub__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for -: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-pub fn mul(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_mul(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_mul(a, b);
-        }
-        if is_float_pair(a, b) {
-            return float_mul(a, b);
-        }
-        if is_str(a) && is_int(b) {
-            return str_repeat(a, b);
-        }
-        if is_int(a) && is_str(b) {
-            return str_repeat(b, a);
-        }
-        // list * int
-        if is_list(a) && is_int(b) {
-            return list_repeat(a, b);
-        }
-        if is_int(a) && is_list(b) {
-            return list_repeat(b, a);
-        }
-        // tuple * int
-        if is_tuple(a) && is_int(b) {
-            let n = w_int_get_value(b).max(0) as usize;
-            let len = w_tuple_len(a);
-            let mut items = Vec::with_capacity(n * len);
-            for _ in 0..n {
-                for i in 0..len {
-                    if let Some(item) = w_tuple_getitem(a, i as i64) {
-                        items.push(item);
-                    }
-                }
-            }
-            return Ok(w_tuple_new(items));
-        }
-        if is_int(a) && is_tuple(b) {
-            return mul(b, a);
-        }
-        // bytes/bytearray * int — bytesobject.py descr_mul / bytearrayobject.py descr_mul
-        if pyre_object::bytesobject::is_bytes_like(a) && is_int(b) {
-            let data = pyre_object::bytesobject::bytes_like_data(a);
-            let n = w_int_get_value(b).max(0) as usize;
-            let mut buf = Vec::with_capacity(data.len() * n);
-            for _ in 0..n {
-                buf.extend_from_slice(data);
-            }
-            return Ok(if pyre_object::bytesobject::is_bytes(a) {
-                pyre_object::bytesobject::w_bytes_from_bytes(&buf)
-            } else {
-                pyre_object::bytearrayobject::w_bytearray_from_bytes(&buf)
-            });
-        }
-        if is_int(a) && pyre_object::bytesobject::is_bytes_like(b) {
-            return mul(b, a);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__mul__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for *: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-pub fn floordiv(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_floordiv(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_floordiv(a, b);
-        }
-        if is_float_pair(a, b) {
-            return float_floordiv(a, b);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__floordiv__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for //: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-pub fn mod_(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_mod(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_mod(a, b);
-        }
-        if is_float_pair(a, b) {
-            return float_mod(a, b);
-        }
-        // str % args — PyPy: unicodeobject.py mod__String_ANY
-        if is_str(a) {
-            return str_format_percent(a, b);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__mod__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for %: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// `str % args` — printf-style string formatting.
-/// PyPy: unicodeobject.py mod__String_ANY → formatting.py
-unsafe fn str_format_percent(fmt: PyObjectRef, args: PyObjectRef) -> PyResult {
-    let fmt_str = w_str_get_value(fmt);
-    // `formatting.py:39-46 StringFormatter.__init__` — when `args`
-    // is a tuple, `values_w` is the unpacked positional list; for
-    // any other shape (mapping, single value), `values_w` stays
-    // None and `checkconsumed` is skipped.  Pyre tracks the same
-    // distinction via `args_is_tuple` so the trailing surplus-args
-    // check at the end fires only for tuple input — without it,
-    // `"%s"`-only formats against a mapping that exposes
-    // `__getitem__` would always trip "not all arguments converted"
-    // even when the mapping was the (single) intended value.
-    let args_is_tuple = is_tuple(args);
-    let arg_list: Vec<PyObjectRef> = if args_is_tuple {
-        let n = w_tuple_len(args);
-        (0..n)
-            .filter_map(|i| w_tuple_getitem(args, i as i64))
-            .collect()
-    } else {
-        vec![args]
-    };
-
-    let mut result = String::new();
-    let mut arg_idx = 0;
-    let bytes = fmt_str.as_bytes();
-    let mut i = 0;
-    let take_next_arg =
-        |arg_idx: &mut usize, arg_list: &[PyObjectRef]| -> Result<PyObjectRef, PyError> {
-            // `formatting.py:574-582 nextinputvalue` — surfaces the
-            // canonical "not enough arguments" message when the
-            // positional pool runs dry.
-            if *arg_idx >= arg_list.len() {
-                return Err(PyError::type_error(
-                    "not enough arguments for format string",
-                ));
-            }
-            let a = arg_list[*arg_idx];
-            *arg_idx += 1;
-            Ok(a)
-        };
-    while i < bytes.len() {
-        if bytes[i] == b'%' && i + 1 < bytes.len() {
-            i += 1;
-            // Named format: %(name)s — `formatting.py:174-195
-            // getmappingkey` scans up to the *balanced* closing
-            // `)` (paren-depth counting so `%(a(b)c)s` parses key
-            // `a(b)c`), then `space.getitem(args, w_key)` looks up
-            // the mapping value (mapping-like object, not just
-            // exact dict).
-            let named_arg = if i < bytes.len() && bytes[i] == b'(' {
-                i += 1; // skip the opening '('
-                let key_start = i;
-                let mut pcount: usize = 1;
-                while i < bytes.len() {
-                    let c = bytes[i];
-                    if c == b')' {
-                        pcount -= 1;
-                        if pcount == 0 {
-                            break;
-                        }
-                    } else if c == b'(' {
-                        pcount += 1;
-                    }
-                    i += 1;
-                }
-                if i >= bytes.len() {
-                    // `formatting.py:184-186 incomplete format key` —
-                    // ran off the end of the format string without
-                    // the matching `)`.
-                    return Err(PyError::new(
-                        PyErrorKind::ValueError,
-                        "incomplete format key".to_string(),
-                    ));
-                }
-                let key = String::from_utf8_lossy(&bytes[key_start..i]).into_owned();
-                i += 1; // skip the closing ')'
-                // Fast path for exact dict: avoid building a W_StrObject
-                // when we can probe the dict storage directly.
-                if is_dict(args) {
-                    w_dict_getitem_str(args, &key)
-                } else {
-                    let w_key = pyre_object::w_str_new(&key);
-                    Some(getitem(args, w_key)?)
-                }
-            } else {
-                None
-            };
-            // `pypy/objspace/std/formatting.py StringFormatter._parse_spec`
-            // — flags (`-`, `+`, ` `, `0`, `#`), width digits, optional
-            // `.precision` digits, then conversion type.  pyre handles
-            // the common subset; the asterisk (`*`) form is left to a
-            // future round.
-            let mut left_align = false;
-            let mut zero_pad = false;
-            let mut explicit_sign = false;
-            let mut blank_sign = false;
-            let mut alt_form = false;
-            while i < bytes.len() {
-                match bytes[i] {
-                    b'-' => left_align = true,
-                    b'0' => zero_pad = true,
-                    b'+' => explicit_sign = true,
-                    b' ' => blank_sign = true,
-                    // `formatting.py:240-242` — `#` selects the
-                    // alternate form (0x/0o/0b prefix on integers).
-                    b'#' => alt_form = true,
-                    _ => break,
-                }
-                i += 1;
-            }
-            // `formatting.py:266-274 peel_num` — `*` reads width /
-            // precision from the next positional input.
-            let mut width = 0usize;
-            if i < bytes.len() && bytes[i] == b'*' {
-                i += 1;
-                let star_arg = take_next_arg(&mut arg_idx, &arg_list)?;
-                if !is_int_like(star_arg) {
-                    return Err(PyError::type_error("* wants int"));
-                }
-                let val = int_value(star_arg);
-                if val < 0 {
-                    // Negative star width acts like `-` flag with abs(width).
-                    left_align = true;
-                    width = (-val) as usize;
-                } else {
-                    width = val as usize;
-                }
-            } else {
-                while i < bytes.len() && bytes[i].is_ascii_digit() {
-                    width = width * 10 + (bytes[i] - b'0') as usize;
-                    i += 1;
-                }
-            }
-            let mut precision: Option<usize> = None;
-            if i < bytes.len() && bytes[i] == b'.' {
-                i += 1;
-                if i < bytes.len() && bytes[i] == b'*' {
-                    i += 1;
-                    let star_arg = take_next_arg(&mut arg_idx, &arg_list)?;
-                    if !is_int_like(star_arg) {
-                        return Err(PyError::type_error("* wants int"));
-                    }
-                    let v = int_value(star_arg);
-                    precision = Some(if v < 0 { 0 } else { v as usize });
-                } else {
-                    let mut p = 0usize;
-                    while i < bytes.len() && bytes[i].is_ascii_digit() {
-                        p = p * 10 + (bytes[i] - b'0') as usize;
-                        i += 1;
-                    }
-                    precision = Some(p);
-                }
-            }
-            if i >= bytes.len() {
-                break;
-            }
-            let spec = bytes[i] as char;
-            i += 1;
-            if spec == '%' {
-                result.push('%');
-                continue;
-            }
-            let arg = if let Some(na) = named_arg {
-                na
-            } else {
-                take_next_arg(&mut arg_idx, &arg_list)?
-            };
-            // Helper closure: pad-and-emit body string respecting
-            // width/left-align (zero-pad only matters for numeric paths
-            // that build their own body with sign awareness).
-            let pad = |body: String| -> String {
-                if body.chars().count() >= width {
-                    return body;
-                }
-                let need = width - body.chars().count();
-                let mut s = String::with_capacity(width);
-                if left_align {
-                    s.push_str(&body);
-                    for _ in 0..need {
-                        s.push(' ');
-                    }
-                } else {
-                    for _ in 0..need {
-                        s.push(' ');
-                    }
-                    s.push_str(&body);
-                }
-                s
-            };
-            let sign_prefix = |val: i64| -> &'static str {
-                if val < 0 {
-                    "-"
-                } else if explicit_sign {
-                    "+"
-                } else if blank_sign {
-                    " "
-                } else {
-                    ""
-                }
-            };
-            match spec {
-                's' => {
-                    let mut body = crate::py_str(arg);
-                    if let Some(p) = precision {
-                        body = body.chars().take(p).collect();
-                    }
-                    result.push_str(&pad(body));
-                }
-                'r' => {
-                    let mut body = crate::py_repr(arg);
-                    if let Some(p) = precision {
-                        body = body.chars().take(p).collect();
-                    }
-                    result.push_str(&pad(body));
-                }
-                // `formatting.py fmt_a` (CPython `%a`) — repr() force-
-                // ASCII-encoded.  Pyre's `py_repr` already produces
-                // ASCII-clean output for the types it covers, so the
-                // result matches PyPy `fmt_a` for the supported subset.
-                'a' => {
-                    let mut body = crate::py_repr(arg);
-                    if let Some(p) = precision {
-                        body = body.chars().take(p).collect();
-                    }
-                    result.push_str(&pad(body));
-                }
-                // `formatting.py:549-552 fmt_b` — `%b` is a *bytes-only*
-                // conversion (formats a `bytes`/`bytearray`/`__bytes__`
-                // object); for the unicode formatter PyPy calls
-                // `self.unknown_fmtchar()` which raises ValueError
-                // "unsupported format character 'b' (0x62) at index N".
-                'b' => {
-                    return Err(PyError::new(
-                        PyErrorKind::ValueError,
-                        format!("unsupported format character 'b' (0x62) at index {}", i - 1),
-                    ));
-                }
-                // `formatting.py fmt_u` is documented as a deprecated
-                // alias for `%d` and dispatches to the same body.
-                'd' | 'i' | 'u' | 'x' | 'X' | 'o' => {
-                    // `formatting.py:296-302 fmt_d / fmt_x ...` raises
-                    // TypeError when the argument is not int-like
-                    // (`%X format: an integer is required, not <type>`),
-                    // not silently coerce via str().
-                    if !is_int_like(arg) {
-                        return Err(PyError::type_error(format!(
-                            "%{spec} format: an integer is required, not {}",
-                            object_functionstr_type_name(arg),
-                        )));
-                    }
-                    let val = int_value(arg);
-                    let abs = val.unsigned_abs();
-                    let digits = match spec {
-                        'x' => format!("{abs:x}"),
-                        'X' => format!("{abs:X}"),
-                        'o' => format!("{abs:o}"),
-                        _ => format!("{abs}"),
-                    };
-                    // `formatting.py:240-242` — alt-form adds the
-                    // base prefix (`0x`, `0X`, `0o`) for the matching
-                    // radix; `%d/%i/%u` ignore it.  `%b` is not a
-                    // unicode-formatter conversion (rejected above
-                    // per `:549-552 fmt_b`).
-                    let prefix = if alt_form {
-                        match spec {
-                            'x' => "0x",
-                            'X' => "0X",
-                            'o' => "0o",
-                            _ => "",
-                        }
-                    } else {
-                        ""
-                    };
-                    let sign = sign_prefix(val);
-                    // Sign-aware zero pad (`%05d`, `%-5d` parity):
-                    // `formatting.py:_pad_string` reserves width for
-                    // sign + alt-prefix before padding zeros.
-                    // Left-align disables zero pad.
-                    if zero_pad && !left_align {
-                        let need = width.saturating_sub(sign.len() + prefix.len() + digits.len());
-                        let mut s = String::with_capacity(width);
-                        s.push_str(sign);
-                        s.push_str(prefix);
-                        for _ in 0..need {
-                            s.push('0');
-                        }
-                        s.push_str(&digits);
-                        result.push_str(&s);
-                    } else {
-                        result.push_str(&pad(format!("{sign}{prefix}{digits}")));
-                    }
-                }
-                // `formatting.py fmt_e / fmt_E / fmt_g / fmt_G` —
-                // scientific / general float formatting.  Default
-                // precision is 6 (CPython %e/%g), `%g` strips
-                // trailing zeros and chooses scientific vs fixed
-                // based on exponent magnitude (PyPy delegates to
-                // `rfloat.formatd` with the matching format char).
-                'e' | 'E' | 'g' | 'G' => {
-                    // `formatting.py:303-308 fmt_e / fmt_g ...` raises
-                    // TypeError when the argument is not numeric;
-                    // silently substituting 0.0 for non-numeric input
-                    // hid type bugs at the format site.
-                    let val = if is_float(arg) {
-                        pyre_object::floatobject::w_float_get_value(arg)
-                    } else if is_int_like(arg) {
-                        int_value(arg) as f64
-                    } else {
-                        return Err(PyError::type_error(format!(
-                            "must be real number, not {}",
-                            object_functionstr_type_name(arg),
-                        )));
-                    };
-                    let prec = precision.unwrap_or(6);
-                    let abs = val.abs();
-                    let body = match spec {
-                        'e' => normalise_exponent(&format!("{:.*e}", prec, abs), false),
-                        'E' => normalise_exponent(&format!("{:.*E}", prec, abs), true),
-                        // `%g` precision counts significant digits;
-                        // Rust has no built-in formatter that exactly
-                        // matches CPython's `%g` rules, so emit
-                        // scientific when |v| >= 10^prec or < 1e-4
-                        // (matching the CPython threshold), else
-                        // fixed with (prec - 1 - exponent) decimals
-                        // and trailing zero stripping.
-                        'g' | 'G' => format_g_like(abs, prec, spec == 'G', alt_form),
-                        _ => unreachable!(),
-                    };
-                    let sign = if val.is_sign_negative() && !val.is_nan() {
-                        "-"
-                    } else if explicit_sign {
-                        "+"
-                    } else if blank_sign {
-                        " "
-                    } else {
-                        ""
-                    };
-                    if zero_pad && !left_align {
-                        let need = width.saturating_sub(sign.len() + body.len());
-                        let mut s = String::with_capacity(width);
-                        s.push_str(sign);
-                        for _ in 0..need {
-                            s.push('0');
-                        }
-                        s.push_str(&body);
-                        result.push_str(&s);
-                    } else {
-                        result.push_str(&pad(format!("{sign}{body}")));
-                    }
-                }
-                'f' | 'F' => {
-                    // `formatting.py:303-308 fmt_f` — same TypeError as
-                    // `%e/%g` for non-numeric arguments.
-                    let val = if is_float(arg) {
-                        pyre_object::floatobject::w_float_get_value(arg)
-                    } else if is_int_like(arg) {
-                        int_value(arg) as f64
-                    } else {
-                        return Err(PyError::type_error(format!(
-                            "must be real number, not {}",
-                            object_functionstr_type_name(arg),
-                        )));
-                    };
-                    let prec = precision.unwrap_or(6);
-                    let abs_body = format!("{:.*}", prec, val.abs());
-                    let sign = if val.is_sign_negative() && !val.is_nan() {
-                        "-"
-                    } else if explicit_sign {
-                        "+"
-                    } else if blank_sign {
-                        " "
-                    } else {
-                        ""
-                    };
-                    if zero_pad && !left_align {
-                        let need = width.saturating_sub(sign.len() + abs_body.len());
-                        let mut s = String::with_capacity(width);
-                        s.push_str(sign);
-                        for _ in 0..need {
-                            s.push('0');
-                        }
-                        s.push_str(&abs_body);
-                        result.push_str(&s);
-                    } else {
-                        result.push_str(&pad(format!("{sign}{abs_body}")));
-                    }
-                }
-                'c' => {
-                    // `formatting.py:283-294 fmt_c` parity:
-                    //   - int branch: `if value < 0 or value > 0x10FFFF:
-                    //     OverflowError("%c arg not in range(0x110000)")`
-                    //   - str branch: `if len(s) != 1: TypeError(
-                    //     "%c requires int or single character")`
-                    //   - other types: same TypeError.
-                    if is_int_like(arg) {
-                        let v = int_value(arg);
-                        if v < 0 || v > 0x10FFFF {
-                            return Err(PyError::new(
-                                PyErrorKind::OverflowError,
-                                "%c arg not in range(0x110000)".to_string(),
-                            ));
-                        }
-                        let c = char::from_u32(v as u32).ok_or_else(|| {
-                            PyError::new(
-                                PyErrorKind::OverflowError,
-                                "%c arg not in range(0x110000)".to_string(),
-                            )
-                        })?;
-                        result.push_str(&pad(c.to_string()));
-                    } else if is_str(arg) {
-                        let s = w_str_get_value(arg);
-                        if s.chars().count() != 1 {
-                            return Err(PyError::type_error("%c requires int or single character"));
-                        }
-                        result.push_str(&pad(s.to_string()));
-                    } else {
-                        return Err(PyError::type_error(format!(
-                            "%c requires int or char, not {}",
-                            object_functionstr_type_name(arg),
-                        )));
-                    }
-                }
-                // `formatting.py:328-329 unknown_fmtchar` raises
-                // ValueError on any spec char outside FORMATTER_CHARS.
-                _ => {
-                    return Err(PyError::value_error(format!(
-                        "unsupported format character '{spec}' (0x{:x}) at index {}",
-                        spec as u32,
-                        i.saturating_sub(1)
-                    )));
-                }
-            }
-        } else {
-            result.push(bytes[i] as char);
-            i += 1;
-        }
-    }
-    // `formatting.py:572-580 checkconsumed` — surplus positional
-    // arguments are an error only when the input came as a tuple
-    // (PyPy's `values_w` invariant).  Mapping inputs (dict and any
-    // other shape) skip this check because every replacement field
-    // pulls from `args` by name, not by sequential index.
-    if args_is_tuple && arg_idx < arg_list.len() {
-        return Err(PyError::type_error(
-            "not all arguments converted during string formatting",
-        ));
-    }
-    Ok(w_str_new(&result))
-}
-
-/// `formatting.py fmt_g / fmt_G` parity helper — emits the `%g` body
-/// for an *absolute* float value.  CPython's `%g` rules:
-///   * precision ≤ 0 is treated as 1 (PyPy: `prec = max(prec, 1)`)
-///   * choose scientific when `exp < -4` or `exp >= prec`
-///     (`pypy/objspace/std/formatting.py:120 format_e_g_complex`)
-///   * scientific body uses (`prec - 1`) decimals
-///   * fixed body uses `(prec - 1 - exp)` decimals
-///   * trailing zeros (and the trailing '.') are stripped unless
-///     `#` (alt-form) was set; with alt-form the trailing zeros AND
-///     the dangling '.' survive so the output advertises the full
-///     requested precision.
-pub(crate) fn format_g_like(abs: f64, prec: usize, upper: bool, alt_form: bool) -> String {
-    if abs == 0.0 {
-        return if alt_form {
-            // `formatting.py:120-128` — alt-form `%#g` keeps `prec`
-            // significant digits even for 0.0, so `"%#.4g" % 0`
-            // renders as `"0.000"` (one digit before the dot, then
-            // `prec - 1` zeros after).
-            let after = prec.saturating_sub(1);
-            if after == 0 {
-                "0".to_string()
-            } else {
-                format!("0.{}", "0".repeat(after))
-            }
-        } else {
-            "0".to_string()
-        };
-    }
-    if !abs.is_finite() {
-        return format!("{abs}");
-    }
-    let prec = prec.max(1);
-    let exp = abs.log10().floor() as i32;
-    let use_sci = exp < -4 || exp >= prec as i32;
-    let raw = if use_sci {
-        let rust_exp = format!("{:.*e}", prec - 1, abs);
-        normalise_exponent(&rust_exp, upper)
-    } else {
-        let dec = (prec as i32 - 1 - exp).max(0) as usize;
-        format!("{abs:.dec$}")
-    };
-    if alt_form {
-        // Alt-form preserves trailing zeros and the dangling '.';
-        // ensure a '.' is present even when the natural body has
-        // none (e.g. `"%#g" % 1` → `"1.00000"`).
-        if use_sci {
-            return raw;
-        }
-        return if raw.contains('.') {
-            raw
-        } else {
-            // `prec - 1 - exp` was 0, so the body has no decimals;
-            // re-render with the alt-form '.' suffix + trailing
-            // zeros to advertise the full precision.
-            let after = (prec as i32 - 1 - exp).max(0) as usize;
-            if after == 0 {
-                format!("{raw}.")
-            } else {
-                format!("{raw}.{}", "0".repeat(after))
-            }
-        };
-    }
-    // Strip trailing zeros from the mantissa (and a dangling '.').
-    if use_sci {
-        if let Some(epos) = raw.find(|c: char| c == 'e' || c == 'E') {
-            let (mantissa, exp_part) = raw.split_at(epos);
-            let trimmed = trim_trailing_zeros(mantissa);
-            format!("{trimmed}{exp_part}")
-        } else {
-            raw
-        }
-    } else if raw.contains('.') {
-        trim_trailing_zeros(&raw)
-    } else {
-        raw
-    }
-}
-
-/// Convert Rust's `{:e}` / `{:E}` exponent encoding (no sign, minimal
-/// width — `1.5e3` / `1.5e-3`) to PyPy / CPython `%e`-style
-/// (`1.5e+03` / `1.5e-03` — explicit sign, exponent zero-padded to at
-/// least two digits).  Mirrors `pypy/objspace/std/formatting.py
-/// fmt_e` which delegates to `rfloat.formatd` with the C printf-style
-/// padding rules.
-pub(crate) fn normalise_exponent(raw: &str, upper: bool) -> String {
-    let marker = if upper { 'E' } else { 'e' };
-    let lower_marker = marker.to_ascii_lowercase();
-    let upper_marker = marker.to_ascii_uppercase();
-    let pos = match raw.find([lower_marker, upper_marker].as_ref()) {
-        Some(p) => p,
-        None => return raw.to_string(),
-    };
-    let (mantissa, exp_part) = raw.split_at(pos);
-    let exp_str = &exp_part[1..]; // skip the marker
-    let (sign_char, digits) = if let Some(rest) = exp_str.strip_prefix('-') {
-        ('-', rest)
-    } else if let Some(rest) = exp_str.strip_prefix('+') {
-        ('+', rest)
-    } else {
-        ('+', exp_str)
-    };
-    let padded = if digits.len() < 2 {
-        format!("0{digits}")
-    } else {
-        digits.to_string()
-    };
-    format!("{mantissa}{marker}{sign_char}{padded}")
-}
-
-fn trim_trailing_zeros(body: &str) -> String {
-    if !body.contains('.') {
-        return body.to_string();
-    }
-    let trimmed = body.trim_end_matches('0').trim_end_matches('.');
-    if trimmed.is_empty() {
-        "0".to_string()
-    } else {
-        trimmed.to_string()
-    }
-}
-
-/// True division (`/` operator) — always produces a float result.
-
-pub fn truediv(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        let a_num = is_int(a) || is_float(a) || is_long(a);
-        let b_num = is_int(b) || is_float(b) || is_long(b);
-        if a_num && b_num {
-            return float_truediv(a, b);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__truediv__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for /: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Power operation dispatch (`**` operator).
-
-pub fn pow(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_pow(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_pow(a, b);
-        }
-        if is_float_pair(a, b) {
-            return float_pow_impl(as_float(a), as_float(b));
-        }
-        if let Some(result) = try_instance_binop(a, b, "__pow__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for **: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-// ── descroperation helpers — pypy/objspace/descroperation.py ──────────
-//
-// These helpers implement the standard "forward + reverse with
-// NotImplemented fallback" dispatch that PyPy generates from
-// `_make_binop_impl` / `_make_descr_unaryop`. They live here in
-// `baseobjspace` (not in `builtins`) because they are space-level
-// semantics shared between the builtin module, the weakproxy wrappers,
-// and any future opcode dispatch — every caller needs the same rule
-// or NotImplemented from the forward path silently swallows the
-// reflected operand.
-
-/// `space.lookup(w_obj, dunder)` — descroperation.py.
-pub(crate) unsafe fn lookup_type_special(obj: PyObjectRef, dunder: &str) -> Option<PyObjectRef> {
-    crate::typedef::r#type(obj).and_then(|tp| lookup_in_type(tp, dunder))
-}
-
-/// descroperation.py `_binop_impl` — when `type(rhs)` is a proper
-/// subtype of `type(lhs)` and defines the reflected dunder, the
-/// reflected operand is tried first.
-pub(crate) unsafe fn should_try_reverse_first(
-    lhs: PyObjectRef,
-    rhs: PyObjectRef,
-    rdunder: &str,
-) -> bool {
-    let Some(lhs_type) = crate::typedef::r#type(lhs) else {
-        return false;
-    };
-    let Some(rhs_type) = crate::typedef::r#type(rhs) else {
-        return false;
-    };
-    !std::ptr::eq(lhs_type, rhs_type)
-        && issubtype_w(rhs_type, lhs_type)
-        && lookup_in_type(rhs_type, rdunder).is_some()
-}
-
-/// Call a special method and treat NotImplemented as "no result", per
-/// descroperation.py `_check_notimplemented`.
-pub(crate) fn try_call_special(
-    method: PyObjectRef,
-    args: &[PyObjectRef],
-) -> Result<Option<PyObjectRef>, PyError> {
-    let result = crate::call::call_function_impl_result(method, args)?;
-    if unsafe { is_not_implemented(result) } {
-        Ok(None)
-    } else {
-        Ok(Some(result))
-    }
-}
-
-/// descroperation.py `_make_binop_impl` — forward `__op__` then
-/// reflected `__rop__`, with the reflected-first reordering rule when
-/// `type(rhs)` subtypes `type(lhs)`.
-pub(crate) fn try_dispatch_binary_special(
-    lhs: PyObjectRef,
-    rhs: PyObjectRef,
-    dunder: &str,
-    rdunder: &str,
-) -> Result<Option<PyObjectRef>, PyError> {
-    let try_reverse_first = unsafe { should_try_reverse_first(lhs, rhs, rdunder) };
-    if try_reverse_first {
-        if let Some(method) = unsafe { lookup_type_special(rhs, rdunder) } {
-            if let Some(result) = try_call_special(method, &[rhs, lhs])? {
-                return Ok(Some(result));
-            }
-        }
-    }
-    if let Some(method) = unsafe { lookup_type_special(lhs, dunder) } {
-        if let Some(result) = try_call_special(method, &[lhs, rhs])? {
-            return Ok(Some(result));
-        }
-    }
-    if !try_reverse_first {
-        if let Some(method) = unsafe { lookup_type_special(rhs, rdunder) } {
-            if let Some(result) = try_call_special(method, &[rhs, lhs])? {
-                return Ok(Some(result));
-            }
-        }
-    }
-    Ok(None)
-}
-
-/// descroperation.py:399 `def pow(space, w_obj1, w_obj2, w_obj3)` —
-/// the same forward/reverse dance as the binary version but threading
-/// the third (modulo) operand through to both arms.
-pub(crate) fn try_dispatch_ternary_special(
-    lhs: PyObjectRef,
-    rhs: PyObjectRef,
-    third: PyObjectRef,
-    dunder: &str,
-    rdunder: &str,
-) -> Result<Option<PyObjectRef>, PyError> {
-    let try_reverse_first = unsafe { should_try_reverse_first(lhs, rhs, rdunder) };
-    if try_reverse_first {
-        if let Some(method) = unsafe { lookup_type_special(rhs, rdunder) } {
-            if let Some(result) = try_call_special(method, &[rhs, lhs, third])? {
-                return Ok(Some(result));
-            }
-        }
-    }
-    if let Some(method) = unsafe { lookup_type_special(lhs, dunder) } {
-        if let Some(result) = try_call_special(method, &[lhs, rhs, third])? {
-            return Ok(Some(result));
-        }
-    }
-    if !try_reverse_first {
-        if let Some(method) = unsafe { lookup_type_special(rhs, rdunder) } {
-            if let Some(result) = try_call_special(method, &[rhs, lhs, third])? {
-                return Ok(Some(result));
-            }
-        }
-    }
-    Ok(None)
-}
-
-/// `(int|long) ** (int|long) % (int|long)` fast path used by `space.pow`
-/// when a modulus is supplied — longobject.py `int_pow`.
-pub(crate) fn try_int_long_pow_with_modulo(
-    base: PyObjectRef,
-    exp: PyObjectRef,
-    modulus: PyObjectRef,
-) -> Result<Option<PyObjectRef>, PyError> {
-    unsafe {
-        if !is_int_or_long(base) || !is_int_or_long(exp) || !is_int_or_long(modulus) {
-            return Ok(None);
-        }
-
-        let base = crate::builtins::obj_to_bigint(base);
-        let exp = crate::builtins::obj_to_bigint(exp);
-        let modulus = crate::builtins::obj_to_bigint(modulus);
-
-        if bigint_eq(modulus.clone(), BigInt::from(0)) {
-            return Err(PyError::value_error("pow() 3rd argument cannot be 0"));
-        }
-        if bigint_lt(exp.clone(), BigInt::from(0)) {
-            return Err(PyError::type_error(
-                "pow() 2nd argument cannot be negative when 3rd argument specified",
-            ));
-        }
-        if bigint_eq(exp.clone(), BigInt::from(0)) {
-            return Ok(Some(box_bigint_result(bigint_mod(
-                BigInt::from(1),
-                modulus,
-            ))));
-        }
-
-        let negative_modulus = bigint_lt(modulus.clone(), BigInt::from(0));
-        let abs_modulus = if negative_modulus {
-            bigint_neg(modulus.clone())
-        } else {
-            modulus.clone()
-        };
-        let mut result = base.modpow(&exp, &abs_modulus);
-        if negative_modulus && bigint_gt(result.clone(), BigInt::from(0)) {
-            result = bigint_sub(result, abs_modulus);
-        }
-        Ok(Some(box_bigint_result(result)))
-    }
-}
-
-pub(crate) fn box_bigint_result(value: BigInt) -> PyObjectRef {
-    use num_traits::ToPrimitive;
-    if let Some(small) = value.to_i64() {
-        w_int_new(small)
-    } else {
-        w_long_new(value)
-    }
-}
-
-pub(crate) fn binary_builtin_type_error(
-    opname: &str,
-    lhs: PyObjectRef,
-    rhs: PyObjectRef,
-) -> PyError {
-    let lhs_name = unsafe {
-        match crate::typedef::r#type(lhs) {
-            Some(tp) => pyre_object::w_type_get_name(tp).to_string(),
-            None => (*(*lhs).ob_type).name.to_string(),
-        }
-    };
-    let rhs_name = unsafe {
-        match crate::typedef::r#type(rhs) {
-            Some(tp) => pyre_object::w_type_get_name(tp).to_string(),
-            None => (*(*rhs).ob_type).name.to_string(),
-        }
-    };
-    PyError::type_error(format!(
-        "unsupported operand type(s) for {opname}: '{lhs_name}' and '{rhs_name}'"
-    ))
-}
-
 /// pypy/interpreter/baseobjspace.py `issubtype_w` — `cls` is in
 /// `w_type.mro_w`. Uses the cached MRO when present, otherwise
 /// recomputes via `compute_default_mro`.
@@ -1978,50 +222,6 @@ pub fn exception_getclass(w_obj: PyObjectRef) -> PyObjectRef {
 ///       return self.issubtype_w(w_cls1, w_cls2)
 pub unsafe fn exception_issubclass_w(w_cls1: PyObjectRef, w_cls2: PyObjectRef) -> bool {
     unsafe { issubtype_w(w_cls1, w_cls2) }
-}
-
-/// 3-arg `pow(a, b, c)` dispatch — pypy/objspace/descroperation.py:399
-/// `def pow(space, w_obj1, w_obj2, w_obj3)`. Tries the int/long modulus
-/// fast path, then forward `__pow__` and reverse `__rpow__` with the
-/// usual NotImplemented fallback. PyPy's MethodTable lists `pow` with
-/// arity=3 (`('pow', '**', 3, ['__pow__', '__rpow__'])`) so a 3-arg
-/// space op exists for the proxy wrapper to call.
-pub fn pow3(base: PyObjectRef, exp: PyObjectRef, modulus: PyObjectRef) -> PyResult {
-    let base = unwrap_cell(base);
-    let exp = unwrap_cell(exp);
-    let modulus = unwrap_cell(modulus);
-    if unsafe { is_none(modulus) } {
-        return pow(base, exp);
-    }
-    if let Some(result) = try_int_long_pow_with_modulo(base, exp, modulus)? {
-        return Ok(result);
-    }
-    if let Some(result) = try_dispatch_ternary_special(base, exp, modulus, "__pow__", "__rpow__")? {
-        return Ok(result);
-    }
-    Err(binary_builtin_type_error("pow()", base, exp))
-}
-
-/// `divmod(a, b)` dispatch — pypy/interpreter/baseobjspace.py:2159
-/// `('divmod', 'divmod', 2, ['__divmod__', '__rdivmod__'])`. Numeric
-/// fast path then forward + reverse special-method dispatch with the
-/// standard NotImplemented fallback.
-pub fn divmod(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        let lhs_num = is_int(a) || is_long(a) || is_float(a);
-        let rhs_num = is_int(b) || is_long(b) || is_float(b);
-        if lhs_num && rhs_num {
-            let q = floordiv(a, b)?;
-            let r = mod_(a, b)?;
-            return Ok(w_tuple_new(vec![q, r]));
-        }
-    }
-    if let Some(result) = try_dispatch_binary_special(a, b, "__divmod__", "__rdivmod__")? {
-        return Ok(result);
-    }
-    Err(binary_builtin_type_error("divmod()", a, b))
 }
 
 /// abstractinst.py:18-31 `_get_bases(space, w_cls)`.
@@ -2354,629 +554,6 @@ pub fn issubclass(derived: PyObjectRef, classinfo: PyObjectRef) -> Result<bool, 
     }
 }
 
-/// floatobject.py:799-881: `_pow(space, x, y)` parity — raw float power
-/// with Python-correct semantics. Handles NaN/Inf edge cases and raises
-/// ZeroDivisionError / ValueError / OverflowError for domain errors.
-///
-/// Returns raw `f64` (matching RPython which returns `r_float`) so the
-/// JIT fast path can call this directly with raw-float arguments without
-/// any intermediate W_FloatObject allocation. The interpreter wrapper
-/// `float_pow_impl` boxes the result into a W_FloatObject.
-pub fn float_pow_raw(x: f64, y: f64) -> Result<f64, PyError> {
-    // floatobject.py:800-801
-    if y == 2.0 {
-        return Ok(x * x);
-    }
-    // floatobject.py:803-804
-    if y == 0.0 {
-        return Ok(1.0);
-    }
-    // floatobject.py:806-807
-    if x.is_nan() {
-        return Ok(x);
-    }
-    // floatobject.py:809-814
-    if y.is_nan() {
-        return Ok(if x == 1.0 { 1.0 } else { y });
-    }
-    // floatobject.py:815-827
-    if y.is_infinite() {
-        let ax = x.abs();
-        if ax == 1.0 {
-            return Ok(1.0);
-        }
-        return Ok(if (y > 0.0) == (ax > 1.0) {
-            f64::INFINITY
-        } else {
-            0.0
-        });
-    }
-    // floatobject.py:828-842
-    if x.is_infinite() {
-        let y_is_odd = y.abs() % 2.0 == 1.0;
-        return Ok(if y > 0.0 {
-            if y_is_odd { x } else { x.abs() }
-        } else if y_is_odd {
-            f64::copysign(0.0, x)
-        } else {
-            0.0
-        });
-    }
-    // floatobject.py:844-847
-    if x == 0.0 && y < 0.0 {
-        return Err(PyError::zero_division(
-            "0.0 cannot be raised to a negative power",
-        ));
-    }
-    // floatobject.py:849-862
-    let mut negate_result = false;
-    let mut bx = x;
-    if bx < 0.0 {
-        if y.floor() != y {
-            return Err(PyError::value_error(
-                "negative number cannot be raised to a fractional power",
-            ));
-        }
-        bx = -bx;
-        negate_result = y.abs() % 2.0 == 1.0;
-    }
-    // floatobject.py:864-869
-    if bx == 1.0 {
-        return Ok(if negate_result { -1.0 } else { 1.0 });
-    }
-    // floatobject.py:871-877
-    let z = bx.powf(y);
-    if z.is_infinite() && !bx.is_infinite() {
-        return Err(PyError::overflow_error("float power"));
-    }
-    // floatobject.py:879-881
-    Ok(if negate_result { -z } else { z })
-}
-
-/// floatobject.py:562 `W_FloatObject.descr_pow` boxing wrapper over `_pow`.
-/// Calls `float_pow_raw` and boxes the raw result into W_FloatObject.
-fn float_pow_impl(x: f64, y: f64) -> PyResult {
-    use pyre_object::w_float_new;
-    Ok(w_float_new(float_pow_raw(x, y)?))
-}
-
-/// Left shift dispatch (`<<` operator).
-
-pub fn lshift(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_lshift(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_lshift(a, b);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__lshift__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for <<: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Right shift dispatch (`>>` operator).
-
-pub fn rshift(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return int_rshift(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_rshift(a, b);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__rshift__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for >>: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Bitwise AND dispatch (`&` operator).
-
-pub fn and_(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        // boolobject.py:74 W_BoolObject.descr_and — both operands bool
-        // → space.newbool(op(a, b)). MRO ensures this runs before the
-        // W_IntObject.descr_and fallback in int_bitand.
-        if is_bool(a) && is_bool(b) {
-            return Ok(pyre_object::bool_descr_and(a, b));
-        }
-        if is_int(a) && is_int(b) {
-            return int_bitand(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_bitand(a, b);
-        }
-        // set / frozenset intersection — PyPy: setobject.py W_BaseSetObject.descr_and
-        if pyre_object::is_set_or_frozenset(a) {
-            let other_items = crate::builtins::collect_iterable(b)?;
-            let probe = pyre_object::w_set_from_items(&other_items);
-            let result: Vec<PyObjectRef> = pyre_object::w_set_items(a)
-                .into_iter()
-                .filter(|&item| pyre_object::w_set_contains(probe, item))
-                .collect();
-            return Ok(if pyre_object::is_frozenset(a) {
-                pyre_object::w_frozenset_from_items(&result)
-            } else {
-                pyre_object::w_set_from_items(&result)
-            });
-        }
-        if let Some(result) = try_instance_binop(a, b, "__and__") {
-            return result;
-        }
-        // Built-in W_Root types (dict_view, …) expose `__and__` /
-        // `__rand__` through their typedef but are not is_instance
-        // — fall back to typedef-driven MRO dispatch before TypeError.
-        if let Some(result) = try_typedef_binop(a, b, "__and__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for &: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Check if an object can participate in `X | Y` union syntax.
-///
-/// PyPy equivalent: _unionable() in _pypy_generic_alias.py
-#[inline]
-fn unionable(obj: PyObjectRef) -> bool {
-    unsafe { is_none(obj) || is_type(obj) || pyre_object::is_union(obj) }
-}
-
-/// Bitwise OR dispatch (`|` operator).
-
-pub fn or_(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    // `pypy/objspace/std/dictproxyobject.py:51 descr_or` /
-    // `pypy/objspace/std/dictproxyobject.py:60 descr_ror` —
-    // mappingproxy `|` dispatches by copying the proxy's wrapped
-    // mapping then `update`-ing with the other operand.  Pre-unwrap
-    // each side so the dict-arm below sees plain dicts and produces
-    // the same merge result.  The proxy-on-rhs case mirrors
-    // `descr_ror` (proxy wraps the rhs operand inside `__or__`).
-    let a = unsafe {
-        if pyre_object::is_dict_proxy(a) {
-            pyre_object::w_dict_proxy_get_mapping(a)
-        } else {
-            a
-        }
-    };
-    let b = unsafe {
-        if pyre_object::is_dict_proxy(b) {
-            pyre_object::w_dict_proxy_get_mapping(b)
-        } else {
-            b
-        }
-    };
-    unsafe {
-        // boolobject.py:75 W_BoolObject.descr_or — both operands bool
-        // → space.newbool(op(a, b)).
-        if is_bool(a) && is_bool(b) {
-            return Ok(pyre_object::bool_descr_or(a, b));
-        }
-        if is_int(a) && is_int(b) {
-            return int_bitor(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_bitor(a, b);
-        }
-        // set / frozenset union — PyPy: setobject.py W_BaseSetObject.descr_or
-        if pyre_object::is_set_or_frozenset(a) {
-            let mut items = pyre_object::w_set_items(a);
-            for item in crate::builtins::collect_iterable(b)? {
-                items.push(item);
-            }
-            return Ok(if pyre_object::is_frozenset(a) {
-                pyre_object::w_frozenset_from_items(&items)
-            } else {
-                pyre_object::w_set_from_items(&items)
-            });
-        }
-        // dict | dict — PEP 584 merge. PyPy: dictmultiobject.py descr_or.
-        // Returns a new dict built from `a`'s items, then updated with `b`'s.
-        if pyre_object::is_dict(a) && pyre_object::is_dict(b) {
-            let new_dict = pyre_object::w_dict_new();
-            for (k, v) in pyre_object::w_dict_items(a) {
-                pyre_object::w_dict_store(new_dict, k, v);
-            }
-            for (k, v) in pyre_object::w_dict_items(b) {
-                pyre_object::w_dict_store(new_dict, k, v);
-            }
-            return Ok(new_dict);
-        }
-        if let Some(result) = try_instance_binop(a, b, "__or__") {
-            return result;
-        }
-        // type | type — PEP 604 union types (Python 3.10+)
-        // PyPy: typeobject.py descr_or → _pypy_generic_alias._create_union
-        if unionable(a) && unionable(b) {
-            return Ok(pyre_object::w_union_new(a, b));
-        }
-        if let Some(result) = try_instance_binop(a, b, "__ror__") {
-            return result;
-        }
-        // Built-in W_Root types (dict_view, …) expose `__or__` /
-        // `__ror__` through their typedef.
-        if let Some(result) = try_typedef_binop(a, b, "__or__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for |: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Bitwise XOR dispatch (`^` operator).
-
-pub fn xor(a: PyObjectRef, b: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_bool(a) && is_bool(b) {
-            return Ok(pyre_object::bool_descr_xor(a, b));
-        }
-        if is_int(a) && is_int(b) {
-            return int_bitxor(a, b);
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            return long_bitxor(a, b);
-        }
-        // set / frozenset symmetric difference — `pypy/objspace/std/
-        // setobject.py W_BaseSetObject.descr_xor`.  Mirrors `and_`'s
-        // intersection arm: walk both sides, keep elements present in
-        // exactly one set.  Result type follows the left operand
-        // (frozenset stays frozenset).
-        if pyre_object::is_set_or_frozenset(a) && pyre_object::is_set_or_frozenset(b) {
-            let mut out: Vec<PyObjectRef> = pyre_object::w_set_items(a)
-                .into_iter()
-                .filter(|&item| !pyre_object::w_set_contains(b, item))
-                .collect();
-            for item in pyre_object::w_set_items(b) {
-                if !pyre_object::w_set_contains(a, item) {
-                    out.push(item);
-                }
-            }
-            return Ok(if pyre_object::is_frozenset(a) {
-                pyre_object::w_frozenset_from_items(&out)
-            } else {
-                pyre_object::w_set_from_items(&out)
-            });
-        }
-        if let Some(result) = try_instance_binop(a, b, "__xor__") {
-            return result;
-        }
-        // Built-in W_Root types (dict_view, …) expose `__xor__` /
-        // `__rxor__` through their typedef.
-        if let Some(result) = try_typedef_binop(a, b, "__xor__") {
-            return result;
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "unsupported operand type(s) for ^: '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Comparison operation dispatch.
-
-pub fn compare(a: PyObjectRef, b: PyObjectRef, op: CompareOp) -> PyResult {
-    let a = unwrap_cell(a);
-    let b = unwrap_cell(b);
-    unsafe {
-        if is_int_like(a) && is_int_like(b) {
-            return match op {
-                CompareOp::Lt => int_lt(a, b),
-                CompareOp::Le => int_le(a, b),
-                CompareOp::Gt => int_gt(a, b),
-                CompareOp::Ge => int_ge(a, b),
-                CompareOp::Eq => int_eq(a, b),
-                CompareOp::Ne => int_ne(a, b),
-            };
-        }
-        if is_int_or_long(a) && is_int_or_long(b) {
-            let va = as_bigint(a);
-            let vb = as_bigint(b);
-            return Ok(w_bool_from(match op {
-                CompareOp::Lt => va < vb,
-                CompareOp::Le => va <= vb,
-                CompareOp::Gt => va > vb,
-                CompareOp::Ge => va >= vb,
-                CompareOp::Eq => va == vb,
-                CompareOp::Ne => va != vb,
-            }));
-        }
-        if is_float_pair(a, b) {
-            return match op {
-                CompareOp::Lt => float_lt(a, b),
-                CompareOp::Le => float_le(a, b),
-                CompareOp::Gt => float_gt(a, b),
-                CompareOp::Ge => float_ge(a, b),
-                CompareOp::Eq => float_eq(a, b),
-                CompareOp::Ne => float_ne(a, b),
-            };
-        }
-        if is_str(a) && is_str(b) {
-            let sa = w_str_get_value(a);
-            let sb = w_str_get_value(b);
-            return Ok(w_bool_from(match op {
-                CompareOp::Lt => sa < sb,
-                CompareOp::Le => sa <= sb,
-                CompareOp::Gt => sa > sb,
-                CompareOp::Ge => sa >= sb,
-                CompareOp::Eq => sa == sb,
-                CompareOp::Ne => sa != sb,
-            }));
-        }
-        // Tuple lexicographic comparison — PyPy: tupleobject.py descr_lt / _eq / etc.
-        if is_tuple(a) && is_tuple(b) {
-            let la = w_tuple_len(a);
-            let lb = w_tuple_len(b);
-            let min_len = la.min(lb);
-            for i in 0..min_len {
-                let ea = w_tuple_getitem(a, i as i64).unwrap_or(PY_NULL);
-                let eb = w_tuple_getitem(b, i as i64).unwrap_or(PY_NULL);
-                let eq = match compare(ea, eb, CompareOp::Eq) {
-                    Ok(r) => is_true(r),
-                    Err(_) => false,
-                };
-                if !eq {
-                    return compare(ea, eb, op);
-                }
-            }
-            return Ok(w_bool_from(match op {
-                CompareOp::Lt => la < lb,
-                CompareOp::Le => la <= lb,
-                CompareOp::Gt => la > lb,
-                CompareOp::Ge => la >= lb,
-                CompareOp::Eq => la == lb,
-                CompareOp::Ne => la != lb,
-            }));
-        }
-        // dict equality — `pypy/objspace/std/dictobject.py
-        // W_DictMultiObject.descr_eq` is order-independent: same length
-        // AND each key-value pair in `a` exists with equal value in `b`.
-        // CPython only defines == / != for dicts (no ordering), so we
-        // restrict to those ops; other ops fall through to the dunder
-        // dispatch which currently raises TypeError, matching the
-        // unimplemented `__lt__` etc. on plain dict.
-        if is_dict(a) && is_dict(b) && matches!(op, CompareOp::Eq | CompareOp::Ne) {
-            let la = pyre_object::w_dict_len(a);
-            let lb = pyre_object::w_dict_len(b);
-            let mut equal = la == lb;
-            if equal {
-                for (k, v) in pyre_object::w_dict_items(a) {
-                    match pyre_object::w_dict_lookup(b, k) {
-                        Some(other_v) => {
-                            let same = compare(v, other_v, CompareOp::Eq)
-                                .map(|r| is_true(r))
-                                .unwrap_or(false);
-                            if !same {
-                                equal = false;
-                                break;
-                            }
-                        }
-                        None => {
-                            equal = false;
-                            break;
-                        }
-                    }
-                }
-            }
-            return Ok(w_bool_from(match op {
-                CompareOp::Eq => equal,
-                CompareOp::Ne => !equal,
-                _ => unreachable!(),
-            }));
-        }
-        // `dictmultiobject.py:1619-1623 _is_set_like` parity — when
-        // one side is a set/frozenset and the other is a set-like
-        // dict_view (Keys / Items), the comparison reduces to the
-        // set-set arm with the dict_view materialised through its
-        // snapshot.  Without this arm, `set == d.keys()` would fall
-        // through to `object.__eq__`'s identity check and return
-        // False even when the contents match.
-        if (pyre_object::is_set_or_frozenset(a) || pyre_object::is_set_or_frozenset(b))
-            && (pyre_object::dictviewobject::is_dict_view(a)
-                || pyre_object::dictviewobject::is_dict_view(b))
-        {
-            let view_set_like = |obj: PyObjectRef| -> bool {
-                if pyre_object::is_set_or_frozenset(obj) {
-                    return true;
-                }
-                if pyre_object::dictviewobject::is_dict_view(obj) {
-                    let kind = pyre_object::dictviewobject::w_dict_view_get_kind(obj);
-                    return matches!(
-                        kind,
-                        pyre_object::dictviewobject::DictViewKind::Keys
-                            | pyre_object::dictviewobject::DictViewKind::Items
-                    );
-                }
-                false
-            };
-            if view_set_like(a) && view_set_like(b) {
-                let a_items = if pyre_object::is_set_or_frozenset(a) {
-                    pyre_object::w_set_items(a)
-                } else {
-                    crate::type_methods::dict_view_snapshot(a)
-                };
-                let b_items = if pyre_object::is_set_or_frozenset(b) {
-                    pyre_object::w_set_items(b)
-                } else {
-                    crate::type_methods::dict_view_snapshot(b)
-                };
-                let a_set = pyre_object::w_set_from_items(&a_items);
-                let b_set = pyre_object::w_set_from_items(&b_items);
-                let la = pyre_object::w_set_len(a_set);
-                let lb = pyre_object::w_set_len(b_set);
-                let a_subset_b = || {
-                    pyre_object::w_set_items(a_set)
-                        .into_iter()
-                        .all(|item| pyre_object::w_set_contains(b_set, item))
-                };
-                let b_subset_a = || {
-                    pyre_object::w_set_items(b_set)
-                        .into_iter()
-                        .all(|item| pyre_object::w_set_contains(a_set, item))
-                };
-                return Ok(w_bool_from(match op {
-                    CompareOp::Eq => la == lb && a_subset_b(),
-                    CompareOp::Ne => la != lb || !a_subset_b(),
-                    CompareOp::Le => la <= lb && a_subset_b(),
-                    CompareOp::Lt => la < lb && a_subset_b(),
-                    CompareOp::Ge => la >= lb && b_subset_a(),
-                    CompareOp::Gt => la > lb && b_subset_a(),
-                }));
-            }
-        }
-        // set / frozenset comparison — subset / superset / equality.
-        // PyPy: setobject.py W_BaseSetObject.descr_eq, descr_le, descr_lt
-        if pyre_object::is_set_or_frozenset(a) && pyre_object::is_set_or_frozenset(b) {
-            let la = pyre_object::w_set_len(a);
-            let lb = pyre_object::w_set_len(b);
-            let a_subset_b = || {
-                pyre_object::w_set_items(a)
-                    .into_iter()
-                    .all(|item| pyre_object::w_set_contains(b, item))
-            };
-            let b_subset_a = || {
-                pyre_object::w_set_items(b)
-                    .into_iter()
-                    .all(|item| pyre_object::w_set_contains(a, item))
-            };
-            return Ok(w_bool_from(match op {
-                CompareOp::Eq => la == lb && a_subset_b(),
-                CompareOp::Ne => la != lb || !a_subset_b(),
-                CompareOp::Le => la <= lb && a_subset_b(),
-                CompareOp::Lt => la < lb && a_subset_b(),
-                CompareOp::Ge => la >= lb && b_subset_a(),
-                CompareOp::Gt => la > lb && b_subset_a(),
-            }));
-        }
-        // List lexicographic comparison — same logic as tuple.
-        if is_list(a) && is_list(b) {
-            let la = pyre_object::w_list_len(a);
-            let lb = pyre_object::w_list_len(b);
-            let min_len = la.min(lb);
-            for i in 0..min_len {
-                let ea = pyre_object::w_list_getitem(a, i as i64).unwrap_or(PY_NULL);
-                let eb = pyre_object::w_list_getitem(b, i as i64).unwrap_or(PY_NULL);
-                let eq = match compare(ea, eb, CompareOp::Eq) {
-                    Ok(r) => is_true(r),
-                    Err(_) => false,
-                };
-                if !eq {
-                    return compare(ea, eb, op);
-                }
-            }
-            return Ok(w_bool_from(match op {
-                CompareOp::Lt => la < lb,
-                CompareOp::Le => la <= lb,
-                CompareOp::Gt => la > lb,
-                CompareOp::Ge => la >= lb,
-                CompareOp::Eq => la == lb,
-                CompareOp::Ne => la != lb,
-            }));
-        }
-        // Instance dunder dispatch for comparison
-        let dunder = match op {
-            CompareOp::Lt => "__lt__",
-            CompareOp::Le => "__le__",
-            CompareOp::Gt => "__gt__",
-            CompareOp::Ge => "__ge__",
-            CompareOp::Eq => "__eq__",
-            CompareOp::Ne => "__ne__",
-        };
-        if let Some(result) = try_instance_binop(a, b, dunder) {
-            return result;
-        }
-        // `dictmultiobject.py:1628-1656 SetLikeDictView` — dict_keys
-        // / dict_items expose `__eq__` / `__ne__` / `__lt__` / etc.
-        // through the typedef.  `try_instance_binop` only fires for
-        // is_instance-shaped objects, so dict views (a separate
-        // W_Root type) need an explicit MRO-driven dispatch here.
-        // Reflected: if RHS is a dict view, try `b.dunder(a)` —
-        // PyPy's `_is_set_like(other)` short-circuits the LHS-side
-        // descr_eq when the other is set-like, so the reflected call
-        // path is the one that succeeds for `set == d.keys()`.
-        if let Some(a_type) = crate::typedef::r#type(a) {
-            if let Some(method) = lookup_in_type_where(a_type, dunder) {
-                if !crate::is_function(method) {
-                    if let Ok(result) = crate::call::call_function_impl_result(method, &[a, b]) {
-                        if !is_not_implemented(result) {
-                            return Ok(result);
-                        }
-                    }
-                }
-                if crate::is_function(method) {
-                    if let Ok(result) = crate::call::call_function_impl_result(method, &[a, b]) {
-                        if !is_not_implemented(result) {
-                            return Ok(result);
-                        }
-                    }
-                }
-            }
-        }
-        if let Some(rdunder) = reverse_dunder(dunder) {
-            if let Some(b_type) = crate::typedef::r#type(b) {
-                if let Some(method) = lookup_in_type_where(b_type, rdunder) {
-                    if let Ok(result) = crate::call::call_function_impl_result(method, &[b, a]) {
-                        if !is_not_implemented(result) {
-                            return Ok(result);
-                        }
-                    }
-                }
-            }
-        }
-        // Identity comparison fallback for == and !=
-        if matches!(op, CompareOp::Eq) {
-            return Ok(w_bool_from(std::ptr::eq(a, b)));
-        }
-        if matches!(op, CompareOp::Ne) {
-            return Ok(w_bool_from(!std::ptr::eq(a, b)));
-        }
-        let a_name = (*(*a).ob_type).name;
-        let b_name = (*(*b).ob_type).name;
-        Err(PyError::type_error(format!(
-            "'{op:?}' not supported between instances of '{a_name}' and '{b_name}'"
-        )))
-    }
-}
-
-/// Comparison operator enum (mirrors RustPython's ComparisonOperator).
-#[derive(Debug, Clone, Copy)]
-pub enum CompareOp {
-    Lt,
-    Le,
-    Gt,
-    Ge,
-    Eq,
-    Ne,
-}
-
 /// Test if an object is truthy (for branch conditions).
 ///
 /// Python truthiness rules:
@@ -3073,90 +650,6 @@ pub fn is_true(obj: PyObjectRef) -> bool {
             }
         }
         true // default: objects are truthy
-    }
-}
-
-/// Unary positive (`+a`).
-
-pub fn pos(a: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    unsafe {
-        if is_int(a) || is_bool(a) {
-            return Ok(w_int_new(int_value(a)));
-        }
-        if is_long(a) {
-            return Ok(bigint_result(w_long_get_value(a).clone()));
-        }
-        if is_float(a) {
-            return Ok(w_float_new(w_float_get_value(a)));
-        }
-        if let Some(result) = try_instance_unaryop(a, "__pos__") {
-            return result;
-        }
-        if a.is_null() {
-            return Err(PyError::type_error(
-                "bad operand type for unary +: 'NoneType'",
-            ));
-        }
-        Err(PyError::type_error(format!(
-            "bad operand type for unary +: '{}'",
-            (*(*a).ob_type).name,
-        )))
-    }
-}
-
-/// Unary negation.
-
-pub fn neg(a: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    unsafe {
-        if is_int(a) || is_bool(a) {
-            let v = int_value(a);
-            return match v.checked_neg() {
-                Some(r) => Ok(w_int_new(r)),
-                None => Ok(w_long_new(bigint_neg(BigInt::from(v)))),
-            };
-        }
-        if is_long(a) {
-            return Ok(bigint_result(bigint_neg(w_long_get_value(a).clone())));
-        }
-        if is_float(a) {
-            return Ok(w_float_new(-w_float_get_value(a)));
-        }
-        // Instance __neg__
-        if let Some(result) = try_instance_unaryop(a, "__neg__") {
-            return result;
-        }
-        if a.is_null() {
-            return Err(PyError::type_error(
-                "bad operand type for unary -: 'NoneType'",
-            ));
-        }
-        Err(PyError::type_error(format!(
-            "bad operand type for unary -: '{}'",
-            (*(*a).ob_type).name,
-        )))
-    }
-}
-
-/// Unary bitwise inversion.
-
-pub fn invert(a: PyObjectRef) -> PyResult {
-    let a = unwrap_cell(a);
-    unsafe {
-        if is_int(a) || is_bool(a) {
-            return Ok(w_int_new(!int_value(a)));
-        }
-        if is_long(a) {
-            return Ok(bigint_result(bigint_invert(w_long_get_value(a).clone())));
-        }
-        if let Some(result) = try_instance_unaryop(a, "__invert__") {
-            return result;
-        }
-        Err(PyError::type_error(format!(
-            "bad operand type for unary ~: '{}'",
-            (*(*a).ob_type).name,
-        )))
     }
 }
 
@@ -5356,7 +2849,7 @@ pub unsafe fn mutated(w_type: PyObjectRef, key: Option<&str>) {
 /// NOTE: PyPy's elidable wrapper (_pure_lookup_where_with_method_cache) takes
 /// a version_tag argument to invalidate on type mutation. Until pyre has
 /// version tags, this raw lookup must NOT be marked elidable.
-unsafe fn lookup_in_type_where(w_type: PyObjectRef, name: &str) -> Option<PyObjectRef> {
+pub(crate) unsafe fn lookup_in_type_where(w_type: PyObjectRef, name: &str) -> Option<PyObjectRef> {
     if w_type.is_null() || !is_type(w_type) {
         return None;
     }
@@ -7531,7 +5024,7 @@ fn object_functionstr_text_w(w_obj: PyObjectRef) -> Result<String, crate::PyErro
     }
 }
 
-fn object_functionstr_type_name(w_obj: PyObjectRef) -> String {
+pub(crate) fn object_functionstr_type_name(w_obj: PyObjectRef) -> String {
     unsafe {
         match crate::typedef::r#type(w_obj) {
             Some(tp) => pyre_object::w_type_get_name(tp).to_string(),
@@ -8344,232 +5837,6 @@ fn normalize_throw_args(w_type: PyObjectRef, w_val: PyObjectRef) -> PyError {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_int_add() {
-        let a = w_int_new(3);
-        let b = w_int_new(4);
-        let result = add(a, b).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 7) };
-    }
-
-    #[test]
-    fn test_int_compare() {
-        let a = w_int_new(5);
-        let b = w_int_new(10);
-        let result = compare(a, b, CompareOp::Lt).unwrap();
-        unsafe { assert!(w_bool_get_value(result)) };
-    }
-
-    #[test]
-    fn test_zero_division() {
-        let a = w_int_new(5);
-        let b = w_int_new(0);
-        assert!(floordiv(a, b).is_err());
-    }
-
-    #[test]
-    fn test_truthiness() {
-        assert!(is_true(w_int_new(1)));
-        assert!(!is_true(w_int_new(0)));
-        assert!(!is_true(w_none()));
-        assert!(is_true(w_bool_from(true)));
-        assert!(!is_true(w_bool_from(false)));
-    }
-
-    #[test]
-    fn test_int_add_overflow() {
-        let a = w_int_new(i64::MAX);
-        let b = w_int_new(1);
-        let result = add(a, b).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(
-                *w_long_get_value(result),
-                BigInt::from(i64::MAX) + BigInt::from(1)
-            );
-        }
-    }
-
-    #[test]
-    fn test_int_sub_overflow() {
-        let a = w_int_new(i64::MIN);
-        let b = w_int_new(1);
-        let result = sub(a, b).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(
-                *w_long_get_value(result),
-                BigInt::from(i64::MIN) - BigInt::from(1)
-            );
-        }
-    }
-
-    #[test]
-    fn test_int_mul_overflow() {
-        let a = w_int_new(i64::MAX);
-        let b = w_int_new(2);
-        let result = mul(a, b).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(
-                *w_long_get_value(result),
-                BigInt::from(i64::MAX) * BigInt::from(2)
-            );
-        }
-    }
-
-    #[test]
-    fn test_long_add() {
-        let a = w_long_new(BigInt::from(i64::MAX) + BigInt::from(1));
-        let b = w_int_new(100);
-        let result = add(a, b).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(
-                *w_long_get_value(result),
-                BigInt::from(i64::MAX) + BigInt::from(101)
-            );
-        }
-    }
-
-    #[test]
-    fn test_long_demote_to_int() {
-        // long + long that fits back in i64 → W_IntObject
-        let a = w_long_new(BigInt::from(i64::MAX) + BigInt::from(1));
-        let b = w_int_new(-1);
-        let result = add(a, b).unwrap();
-        unsafe {
-            assert!(is_int(result));
-            assert_eq!(w_int_get_value(result), i64::MAX);
-        }
-    }
-
-    #[test]
-    fn test_negate_min_int() {
-        let a = w_int_new(i64::MIN);
-        let result = neg(a).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(*w_long_get_value(result), -BigInt::from(i64::MIN));
-        }
-    }
-
-    #[test]
-    fn test_invert_int() {
-        let result = invert(w_int_new(6)).unwrap();
-        unsafe {
-            assert!(is_int(result));
-            assert_eq!(w_int_get_value(result), !6);
-        }
-    }
-
-    #[test]
-    fn test_long_compare() {
-        let a = w_long_new(BigInt::from(i64::MAX) + BigInt::from(1));
-        let b = w_int_new(i64::MAX);
-        let result = compare(a, b, CompareOp::Gt).unwrap();
-        unsafe { assert!(w_bool_get_value(result)) };
-    }
-
-    #[test]
-    fn test_long_truthiness() {
-        assert!(is_true(w_long_new(
-            BigInt::from(i64::MAX) + BigInt::from(1)
-        )));
-        assert!(!is_true(w_long_new(BigInt::from(0))));
-    }
-
-    #[test]
-    fn test_int_pow() {
-        let result = pow(w_int_new(2), w_int_new(10)).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 1024) };
-    }
-
-    #[test]
-    fn test_int_pow_overflow() {
-        let result = pow(w_int_new(2), w_int_new(63)).unwrap();
-        unsafe {
-            // 2^63 overflows i64, should be long
-            assert!(is_long(result));
-            assert_eq!(*w_long_get_value(result), BigInt::from(2).pow(63));
-        }
-    }
-
-    #[test]
-    fn test_int_pow_negative_exponent() {
-        let result = pow(w_int_new(2), w_int_new(-1)).unwrap();
-        unsafe {
-            assert!(is_float(result));
-            assert_eq!(w_float_get_value(result), 0.5);
-        }
-    }
-
-    #[test]
-    fn test_int_lshift() {
-        let result = lshift(w_int_new(1), w_int_new(10)).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 1024) };
-    }
-
-    #[test]
-    fn test_int_lshift_overflow() {
-        let result = lshift(w_int_new(1), w_int_new(64)).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(*w_long_get_value(result), BigInt::from(1) << 64);
-        }
-    }
-
-    #[test]
-    fn test_int_rshift() {
-        let result = rshift(w_int_new(1024), w_int_new(3)).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 128) };
-    }
-
-    #[test]
-    fn test_negative_shift_count() {
-        assert!(lshift(w_int_new(1), w_int_new(-1)).is_err());
-        assert!(rshift(w_int_new(1), w_int_new(-1)).is_err());
-    }
-
-    #[test]
-    fn test_int_bitand() {
-        let result = and_(w_int_new(0xFF), w_int_new(0x0F)).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 0x0F) };
-    }
-
-    #[test]
-    fn test_int_bitor() {
-        let result = or_(w_int_new(0xF0), w_int_new(0x0F)).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 0xFF) };
-    }
-
-    #[test]
-    fn test_int_bitxor() {
-        let result = xor(w_int_new(0xFF), w_int_new(0x0F)).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 0xF0) };
-    }
-
-    #[test]
-    fn test_long_bitand() {
-        let a = w_long_new(BigInt::from(i64::MAX) + BigInt::from(1));
-        let b = w_int_new(0xFF);
-        let result = and_(a, b).unwrap();
-        unsafe { assert_eq!(w_int_get_value(result), 0) };
-    }
-
-    #[test]
-    fn test_invert_long() {
-        let a = w_long_new(BigInt::from(i64::MAX) + BigInt::from(1));
-        let result = invert(a).unwrap();
-        unsafe {
-            assert!(is_long(result));
-            assert_eq!(
-                *w_long_get_value(result),
-                !(BigInt::from(i64::MAX) + BigInt::from(1))
-            );
-        }
-    }
 
     #[test]
     fn test_setattr_getattr() {
