+++ pyre/pyre-interpreter/src/objspace/mod.rs
@@ -1,3 +1,4 @@
 //! pypy/objspace/
 
+pub mod descroperation;
 pub mod std;
