+++ pyre/pyre-interpreter/src/objspace/std/mod.rs
@@ -1,3 +1,4 @@
 //! pypy/objspace/std/
 
+pub mod formatting;
 pub mod mapdict;
